removed log data binding from pause button
Don't remember why I had it there in the first place (probably from
testing) but it was causing logBox entries to flash on pause as it
updated the logBox with no new data. Also removed the version info in
main window title from the last commit. Decided I didn't like it.