Enable/Disable buttons when they're useable or not
As a visual indicator of when clicking a button will have a result, as
there are cases where there will be no result like clicking finish when
it's already finished. Also gave an invisible object to drop button
focus onto on click in the metro theme because the bright white focus on
the buttons shifting around all the time gets very disorienting.