Added log importing, better log saving, manual save
will update the timestamp on the log file for the current session when
saving. Importing a log introduces it to that session as the session
file, so saving after importing will update the timestamp on the
imported file instead of making a new log.  Manual saving allows the
user to save electively as well. Helps make sure logs are saved in the
case of a power outage or anything that would prevent the program from
going through a proper lifecycle.